requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0

# LLM providers (install based on your choice)
openai>=1.0.0
google-generativeai>=0.3.0
ollama>=0.1.0

# Utilities
requests>=2.31.0

# REST API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
